+++ libqvdclient/c/qvdvm.h
@@ -2,7 +2,7 @@
 #define _QVDVM_H
 #include "qvdclient.h"
 
-vm *QvdVmNew(int64_t id, const char *name, const char *state, int64_t blocked);
+vm *QvdVmNew(int id, const char *name, const char *state, int blocked);
 void QvdVmFree(vm *ptr);
 void QvdVmListInit(vmlist *ptr);
 void QvdVmListAppendVm(qvdclient *qvd, vmlist *vmlistptr, vm *vmptr);
